Fixed Image recieve
Old code had an error where it didnt recieve the whole image, added checks to ensure the response is equal to content length if it is specified. also added an if statement to stop the server from trying to store images with no-store specified. This was because the image had no-store and our method of caching isnt really compatible with image files. Might look more into this if I get time